WHAT ---- 1. deleted TestWindow.java file 2. deleted TestWindowController.java file 3. deleted TestWindow.fxml file 4. deleted UpdateView.java file 5. deleted UpdateViewController.java file 6. changed the UpdateView statment to MainView statement in the SystemTrayUtils.java class 7. added MainView.java, MainView.fxml and MainViewController.java files 8. added Status.fxml and StatusController.java files 9. added ClientUpdate.fxml and ClientUpdateController.java files 10. added ServerUpdate.fxml and ServerUpdateController.java files
