+++ src/main/resources/views/ServerUpdate.fxml
@@ -1,0 +1,86 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import com.jfoenix.controls.JFXButton?>
+<?import com.jfoenix.controls.JFXProgressBar?>
+<?import de.jensd.fx.glyphs.fontawesome.FontAwesomeIconView?>
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.ScrollPane?>
+<?import javafx.scene.control.TextArea?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane fx:id="mainAnchorPane" prefHeight="680.0" prefWidth="790.0" stylesheets="@../styles/updateView.css" xmlns="http://javafx.com/javafx/8.0.141" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.idealupdater.utils.ui.controllers.ServerUpdateController">
+   <children>
+      <AnchorPane fx:id="headerAnchorPane" layoutX="163.0" minHeight="-Infinity" prefHeight="70.0" prefWidth="566.0" styleClass="contentPane" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+         <children>
+            <Label fx:id="headerLabel" layoutY="18.0" prefHeight="70.0" style="-fx-font-family: Consolas;" text="Server Update" textFill="WHITE" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+               <font>
+                  <Font name="System Bold" size="25.0" />
+               </font>
+               <padding>
+                  <Insets left="30.0" />
+               </padding>
+            </Label>
+         </children>
+      </AnchorPane>
+      <JFXProgressBar fx:id="progressBar" layoutX="163.0" layoutY="70.0" prefHeight="9.0" prefWidth="661.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="70.0" />
+      <AnchorPane fx:id="updatesAnchorPane" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="50.0" AnchorPane.rightAnchor="50.0" AnchorPane.topAnchor="80.0">
+         <children>
+            <Label layoutY="43.0" text="New Features" AnchorPane.leftAnchor="0.0" AnchorPane.topAnchor="20.0">
+               <font>
+                  <Font name="System Bold" size="20.0" />
+               </font>
+            </Label>
+            <ScrollPane layoutY="83.0" prefHeight="200.0" prefWidth="200.0" style="-fx-background-color: transparent;" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="70.0">
+               <content>
+                  <AnchorPane maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" minHeight="0.0" minWidth="0.0" prefHeight="200.0" prefWidth="546.0">
+                     <children>
+                        <VBox fx:id="featureVBox" prefHeight="182.0" prefWidth="614.0" spacing="5.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0">
+                           <children>
+                              <Label fx:id="versionLabel" text="V1.0.1">
+                                 <font>
+                                    <Font name="System Bold" size="16.0" />
+                                 </font>
+                              </Label>
+                              <Label text=" ➝ new login button">
+                                 <font>
+                                    <Font size="16.0" />
+                                 </font>
+                              </Label>
+                              <Label text=" ➝ finance module">
+                                 <font>
+                                    <Font size="16.0" />
+                                 </font>
+                              </Label>
+                              <Label text=" ➝ tax report">
+                                 <font>
+                                    <Font size="16.0" />
+                                 </font>
+                              </Label>
+                           </children>
+                        </VBox>
+                     </children>
+                  </AnchorPane>
+               </content>
+            </ScrollPane>
+            <JFXButton fx:id="updateBtn" layoutY="304.0" onAction="#performUpdate" prefHeight="44.0" prefWidth="185.0" styleClass="updateBtn" text="UPDATE" AnchorPane.leftAnchor="0.0" AnchorPane.topAnchor="280.0" />
+            <TextArea fx:id="consoleField" editable="false" prefHeight="200.0" styleClass="console" text="Console Output!" AnchorPane.bottomAnchor="10.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="366.0" />
+            <ScrollPane layoutX="10.0" layoutY="93.0" prefHeight="200.0" prefWidth="200.0" style="-fx-background-color: transparent;" visible="false" AnchorPane.bottomAnchor="10.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="366.0">
+               <content>
+                  <AnchorPane maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" minHeight="0.0" minWidth="0.0" prefHeight="200.0" prefWidth="546.0" />
+               </content>
+            </ScrollPane>
+            <JFXButton fx:id="revertBtn" layoutX="10.0" layoutY="314.0" onAction="#performRevert" prefHeight="44.0" styleClass="revertBtn" text="BACK TO OLD VERSION" visible="false" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="280.0">
+               <padding>
+                  <Insets left="20.0" right="20.0" />
+               </padding>
+               <graphic>
+                  <FontAwesomeIconView fill="WHITE" glyphName="ARROW_LEFT" />
+               </graphic>
+            </JFXButton>
+         </children>
+      </AnchorPane>
+   </children>
+</AnchorPane>
