+++ src/main/resources/views/Status.fxml
@@ -1,0 +1,70 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import com.jfoenix.controls.JFXProgressBar?>
+<?import com.jfoenix.controls.JFXToggleButton?>
+<?import de.jensd.fx.glyphs.fontawesome.FontAwesomeIconView?>
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane fx:id="mainAnchorPane" prefHeight="680.0" prefWidth="790.0" stylesheets="@../styles/updateView.css" xmlns="http://javafx.com/javafx/8.0.141" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.idealupdater.utils.ui.controllers.StatusController">
+   <children>
+      <AnchorPane fx:id="headerAnchorPane" layoutX="163.0" minHeight="-Infinity" prefHeight="70.0" prefWidth="566.0" styleClass="contentPane" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+         <children>
+            <Label fx:id="headerLabel" layoutY="18.0" prefHeight="70.0" style="-fx-font-family: Consolas;" text="Client / Server Status" textFill="WHITE" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+               <font>
+                  <Font name="System Bold" size="25.0" />
+               </font>
+               <padding>
+                  <Insets left="30.0" />
+               </padding>
+            </Label>
+         </children>
+      </AnchorPane>
+      <JFXProgressBar fx:id="progressBar" layoutX="163.0" layoutY="70.0" prefHeight="9.0" prefWidth="661.0" visible="false" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="70.0" />
+      <AnchorPane fx:id="statusAnchorPane" layoutX="213.0" layoutY="92.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+         <children>
+            <Label alignment="CENTER" layoutX="57.0" layoutY="129.0" prefWidth="230.0" AnchorPane.leftAnchor="50.0" AnchorPane.topAnchor="0.0">
+               <graphic>
+                  <FontAwesomeIconView fx:id="clientStatusIcon" fill="#787878" glyphName="LAPTOP" size="72.0" />
+               </graphic>
+            </Label>
+            <Label fx:id="clientStatusLabel" alignment="CENTER" layoutX="18.0" layoutY="300.0" text="Client Status: OFF" textFill="#504d4d" AnchorPane.leftAnchor="50.0" AnchorPane.topAnchor="100.0">
+               <font>
+                  <Font name="Consolas Bold" size="21.0" />
+               </font>
+               <padding>
+                  <Insets left="10.0" right="10.0" />
+               </padding>
+            </Label>
+            <Label alignment="CENTER" layoutX="333.0" layoutY="129.0" prefWidth="230.0" AnchorPane.rightAnchor="50.0" AnchorPane.topAnchor="0.0">
+               <graphic>
+                  <FontAwesomeIconView fx:id="serverStatusIcon" fill="#787878" glyphName="SERVER" size="72.0" styleClass="inactive-node-icon" />
+               </graphic>
+            </Label>
+            <Label fx:id="serverStatusLabel" alignment="CENTER" layoutX="28.0" layoutY="300.0" text="Sever Status: OFF" textFill="#504d4d" AnchorPane.rightAnchor="50.0" AnchorPane.topAnchor="100.0">
+               <font>
+                  <Font name="Consolas Bold" size="21.0" />
+               </font>
+               <padding>
+                  <Insets left="10.0" right="10.0" />
+               </padding>
+            </Label>
+            <JFXToggleButton fx:id="clientToggleBtn" size="20.0" text="OFF" toggleColor="#55ab04" unToggleColor="#b21313" unToggleLineColor="#f56d6d" AnchorPane.leftAnchor="50.0" AnchorPane.topAnchor="150.0">
+               <font>
+                  <Font name="System Bold" size="15.0" />
+               </font>
+            </JFXToggleButton>
+            <JFXToggleButton fx:id="serverToggleBtn" size="20.0" text="OFF" toggleColor="#55ab04" unToggleColor="#b21313" unToggleLineColor="#f56d6d" AnchorPane.rightAnchor="50.0" AnchorPane.topAnchor="150.0">
+               <font>
+                  <Font name="System Bold" size="15.0" />
+               </font>
+            </JFXToggleButton>
+         </children>
+         <padding>
+            <Insets top="100.0" />
+         </padding>
+      </AnchorPane>
+   </children>
+</AnchorPane>
