--- src/main/resources/views/TestWindow.fxml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import com.jfoenix.controls.JFXButton?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.141" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.idealupdater.utils.ui.controllers.TestWindowController">
-   <children>
-      <JFXButton fx:id="btn" layoutX="154.0" layoutY="111.0" onAction="#doSomething" prefHeight="44.0" prefWidth="239.0" style="-fx-background-color: #FF382D;" text="Test Button" textFill="WHITE">
-         <font>
-            <Font name="System Bold" size="16.0" />
-         </font>
-      </JFXButton>
-   </children>
-</AnchorPane>
